// src/types/changeset_spec.schema.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * A changeset specification, which describes a changeset to be created or an existing changeset to be tracked.
 */
export type ChangesetSpec =
  | {
      /**
       * The GraphQL ID of the repository that contains the existing changeset on the code host.
       */
      baseRepository: string;
      /**
       * The ID that uniquely identifies the existing changeset on the code host
       */
      externalID: string;
    }
  | {
      /**
       * The GraphQL ID of the repository that this changeset spec is proposing to change.
       */
      baseRepository: string;
      /**
       * The full name of the Git ref in the base repository that this changeset is based on (and is proposing to be merged into). This ref must exist on the base repository.
       */
      baseRef: string;
      /**
       * The base revision this changeset is based on. It is the latest commit in baseRef at the time when the changeset spec was created.
       */
      baseRev: string;
      /**
       * The GraphQL ID of the repository that contains the branch with this changeset's changes. Fork repositories and cross-repository changesets are not yet supported. Therefore, headRepository must be equal to baseRepository.
       */
      headRepository: string;
      /**
       * The full name of the Git ref that holds the changes proposed by this changeset. This ref will be created or updated with the commits.
       */
      headRef: string;
      /**
       * The title of the changeset on the code host.
       */
      title: string;
      /**
       * The body (description) of the changeset on the code host.
       */
      body: string;
      /**
       * The Git commits with the proposed changes. These commits are pushed to the head ref.
       */
      commits: [GitCommitDescription];
      /**
       * Whether to publish the changeset. An unpublished changeset can be previewed on Sourcegraph by any person who can view the campaign, but its commit, branch, and pull request aren't created on the code host. A published changeset results in a commit, branch, and pull request being created on the code host.
       */
      published: boolean;
    };

/**
 * The Git commit to create with the changes.
 */
export interface GitCommitDescription {
  /**
   * The Git commit message.
   */
  message: string;
  /**
   * The commit diff (in unified diff format).
   */
  diff: string;
  /**
   * The Git commit author name.
   */
  authorName: string;
  /**
   * The Git commit author email.
   */
  authorEmail: string;
}